% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_functions.R
\name{dmsig}
\alias{dmsig}
\title{Density for the Mixture of an Inverse Gaussian and a Uniform
Distribution}
\usage{
dmsig(t, kappa, xi, tau, lambda, alpha = min(t), beta = max(t), ln = F,
  summation = F, sep = F)
}
\arguments{
\item{t}{vector of response times (t > 0).}

\item{kappa}{a scalar threshold value (kappa > 0).}

\item{xi}{a scalar rate of evidence accumulation (xi > 0).}

\item{tau}{a scalar residual latency by which to shift the response times.
( 0 <= tau < min(t) ).}

\item{lambda}{a scalar mixture probability for the inverse gaussian.}

\item{alpha}{a scalar for the lower boundary to use in the uniform density.}

\item{beta}{a scalar for the uppber boundary to use in the uniform density.}

\item{ln}{logical; If TRUE, returns the log of the density.}

\item{summation}{if TRUE, returns the sum of the logs of the
densities.}

\item{sep}{logical; if TRUE, returns a matrix with separate columns for the weighted
likelihoods of the inverse gaussian and the uniform distribution.}
}
\value{
If \code{sep} is TRUE, gives a matrix whose columns are the separate weighted
  densities for the inverse gaussian and uniform distribution, respectively. If
  \code{ln} is FALSE, gives the likelihood, else gives the log-likelihood.
  If both \code{ln} and \code{summation} are TRUE, gives the sum of the log-likelihoods.
}
\description{
The density function for the mixture of a shifted inverse
gaussian and a uniform distribution. The inverse gaussian
distribution is parameterized in terms of a one-boundary
wiener process.
}

