% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_functions.R
\name{dsig}
\alias{dsig}
\title{Density for the Shifted Inverse Gaussian}
\usage{
dsig(t, kappa, xi, tau, ln = F, summation = F)
}
\arguments{
\item{t}{vector of response times (t > 0).}

\item{kappa}{a scalar threshold value (kappa > 0).}

\item{xi}{a scalar rate of evidence accumulation (xi > 0).}

\item{tau}{a scalar residual latency by which to shift the response times.
( 0 <= tau < min(t) ).}

\item{ln}{logical; If TRUE, returns the log of the density.}

\item{summation}{if TRUE, returns the sum of the logs of the
densities.}
}
\value{
If \code{ln} is FALSE, gives the likelihood, else gives the log-likelihood.
  If both \code{ln} and \code{summation} are TRUE, gives the sum of the log-likelihoods.
}
\description{
The density function for a shifted inverse gaussian distribution,
parameterized in terms of a one-boundary wiener process.
}
\details{
The shifted inverse gaussian, when parameterized in terms of brownian motion,
  has density \deqn{ f(t - \tau) = \kappa/\sqrt( 2*\pi*(t-\tau)^3 ) exp( (-.5/(t-\tau))*(\kappa - \xi*(t-\tau))^2 ) }
  where \eqn{t} is a response time, \eqn{\kappa} is a threshold toward which evidence
  accumulates with average rate \eqn{\xi} and a fixed variance of 1, and \eqn{\tau} is a
  residual latency subtracted from the response time.
}

