% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_functions.R
\name{rtclean}
\alias{rtclean}
\title{Function for Trimming Outlier Response Times}
\usage{
rtclean(rt, exclude = 1, lowerBoundary = c(0.5, 0.5, 0, 0.8),
  upperBoundary = c(5, 5, 0.3, 0.95), method = "BFGS", maxit = 5000,
  nRep = 5, alpha = 0.95, cutoff = -2.9957)
}
\arguments{
\item{rt}{a vector of response times.}

\item{exclude}{if \code{exclude} equals 1, any response time that has a relative
probability less that 0.5 under the inverse gaussian distribution is trimmed. If
\code{exclude} equals 2, response times are probabilistically trimmed using the
relative probabilities under the inverse gaussian and uniform distribution.}

\item{lowerBoundary}{a vector defining the lower boundary for the range over which
the starting values are generated.}

\item{upperBoundary}{a vector defining the upper boundary for the range over which
the starting values are generated.}

\item{method}{determines the estimation method used by R's \code{optim} function.}

\item{maxit}{the maximum number of iterations to run in the optimization routine.}

\item{nRep}{the number of times to repeat the optimization routine with new starting
values (higher values lead to a greater chance of avoiding local maxima at the cost
of longer computational times).}

\item{alpha}{the width of the confidence intervals for parameter estimates.}

\item{cutoff}{a cutoff for excluding observations based on their relative
probability under the inverse gaussian (default is .05).}
}
\value{
\code{rtclean} returns an object of class "cleanrt".

The function \code{summary} prints a comparison between the descriptive statistics
for the original and trimmed data sets, \code{anova} reports the Akaike Information
Criterion (AIC) values (corrected for small samples) and their relative weights for
the set of 3 models fit to the data.

An object of class "cleanrt" is a list containing the following components:
\describe{
  \item{\code{rtNew}}{ a vector with the newly trimmed response times.}
  \item{\code{rtOld}}{a vector of the untrimmed response times.}
  \item{\code{M1}}{the \code{optim} results for the inverse gaussian model.}
  \item{\code{M2}}{the \code{optim} results for the shifted inverse gaussian model.}
  \item{\code{M3}}{the \code{optim} results for the mixture model.}
  \item{\code{M3}}{a list with the AICc values and relative weights.}
  \item{\code{ProbIG}}{a vector with the relative probabilities of each observation
    under the shifted inverse gaussian distribution.}
  \item{\code{exclude_1}}{a logical vector for the response times selected for
    exclusion based on having a relative probability of less than 0.5 for the
    shifted inverse gaussian.}
  \item{\code{exclude_2}}{a logical vector for the response times selected for
    exclusion based on probabilistic sampling using the relative probabilities
    for the inverse gaussian and uniform distributions.}
}
}
\description{
Forthcoming
}

